refactor: AppLayout 폴더 이동
