+++ src/styles/utils.scss
@@ -12,3 +12,8 @@ $tag-font: #1d39c4;
 $tag-border: #adc6ff;
 $tag-background: #f0f5ff;
 $preview-background: #f3f3f3;
+
+
+//renew
+$default-white: #ffffff;
+$default-black: #373833;
